vsl-backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
passlib[bcrypt]
python-jose
pydantic
python-dotenv
alembic 
